Fix indentation, add file_get_contents_curl and getDatabasesFromHost
This commit changes the indentation to use tabs a little more
consistently and removes some trailing whitespace.

This commit also adds `file_get_contents_curl` which is a wrapper for
curl that emulates `file_get_contents`. I'm sure I did it for a good
reason but I forget exactly why right now.

The other addition is `getDatabasesFromHost` which adds support for
getting a listing of databases on a CouchDB server that is provided.
This is useful for ensuring that new databases are automatically backed
up from a CouchDB server.