Adding Managed Fields 1.0.1
This release adds support for setIfEmpty, fixes an issue where the
record ID wasn't populated in the field log and adds the ability to
pass in extra rules merge with the field configuration. This version
also fixes a gap where the setting for disabling scripts for the
setManagedFieldValues was missing.